[refactor] Refactor utilities.js to a folder
