+++ utilities/countryTrends.js
@@ -1,6 +1,7 @@
 var _ = require('underscore');
 var fs = require('fs');
 var twitter = require('twitter');
+
 /*************************************************************
 Add Underscore Mixin to sort by keys
 **************************************************************/
@@ -75,24 +76,4 @@ exports.returnTrendsByCountry = function(callback) {
     }
   });
 }
-
-/*************************************************************
-Example of writing to a file
-**************************************************************/
-
-// fs.writeFile('./tweets.txt', tweets, function(err){
-//   if(err){console.log(err)}
-// })
-
-/*************************************************************
-Histogram Function
-**************************************************************/
-
-var histogram = function(ar){
-  return reduce(ar, function(obj, key){
-    obj[key] = obj[key] || 0;
-    obj[key]++;
-    return obj;
-  }, {});
-};
